requirements.txt
fastapi
uvicorn
langgraph
google-genai
psycopg[binary,pool]
langgraph-checkpoint-postgres
python-dotenv
